% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_parts.R
\name{get_centers}
\alias{get_centers}
\title{get_centers}
\usage{
get_centers(parts)
}
\arguments{
\item{parts}{object created by \link[randomviz]{make_parts}}
}
\value{
adds centers as a data frame x and y coordinates and integer group assignment to parts list
}
\description{
get_centers
}
\details{
centers are calculated as the mean between all part x and y coordinates seperately for each group
}
