% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_parts.R
\name{get_colors}
\alias{get_colors}
\title{get_colors}
\usage{
get_colors(obj, color_fun = "random_colors", ...)
}
\arguments{
\item{obj}{containing a list of parts}

\item{fun}{string name of function to generate color for each part}
}
\description{
get_colors
}
