% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_parts.R
\name{get_linear_forces}
\alias{get_linear_forces}
\title{get_centers}
\usage{
get_linear_forces(parts, center, attract = T, increment = 0.5)
}
\arguments{
\item{parts}{object created by \link[randomviz]{make_parts}}

\item{center}{object calculated by \link[randomviz]{get_centers}}

\item{attract}{boolean move parts toward center (TRUE) or away from the center (FALSE)}

\item{increment}{the distance each point is moved as a percent of the total distance to the center. For example increment=0.5 corresponds to a step of half the distance to the center for each step.}
}
\value{
list of parts and data frame of part centers x and y coordinates and integer group assignment
}
\description{
get_centers
}
