% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_parts.R
\name{make_parts}
\alias{make_parts}
\title{make_parts}
\usage{
make_parts(part_num = 10, groups = 3, seed = 1)
}
\arguments{
\item{part_num}{number of points}

\item{groups}{number of groups to assign points to at random}
}
\value{
data frame of part x and y coordinates and integer group assignment
}
\description{
make_parts
}
