% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_parts.R
\name{obj_dims}
\alias{obj_dims}
\title{obj_dims}
\usage{
obj_dims(obj, zoom = NULL)
}
\arguments{
\item{obj}{list of parts}

\item{zoom}{decimal and percent to zoom (e.g. zoom out, 0 to 1; zoom in, -1 to 0)}
}
\value{
adds dims or x and y limits for plotting
}
\description{
obj_dims
}
