% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_parts.R
\name{plot_obj_fun}
\alias{plot_obj_fun}
\title{plot_obj_fun}
\usage{
plot_obj_fun(obj, fun = "points", index = NULL, sleep = 0.1, seed = 123, ...)
}
\arguments{
\item{obj}{list containing list of parts}

\item{fun}{function string to call one of c('points','lines','polygons')}

\item{index}{of which part to plot if NULL all will be printed}

\item{sleep}{pause between plotting parts (E.g. use for animation timing)}

\item{...}{passed to fun}

\item{alpha}{transparency}
}
\description{
plot_obj_fun
}
\details{
plot all parts based on supplied function c('points','lines','polygons')
}
