% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_parts.R
\name{points_linear_attractor}
\alias{points_linear_attractor}
\title{points_linear_attractor}
\usage{
points_linear_attractor(obj, iter = 10, ...)
}
\arguments{
\item{obj}{object created by \link[randomviz]{make_parts} followed by \link[randomviz]{get_centers}}

\item{iter}{integer of iterations to calculate attraction or repulsion based on \link[randomviz]{get_linear_forces}}
}
\value{
list of iteration results for parts
}
\description{
points_linear_attractor
}
