% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_parts.R
\name{random_colors}
\alias{random_colors}
\title{random_colors}
\usage{
random_colors(id = NULL, alpha = 1, seed = 123)
}
\arguments{
\item{id}{lintege specifying index to choose from colors()}

\item{seed}{random seed to scramble colors()}
}
\description{
random_colors
}
