% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_parts.R
\name{rotate_obj_angles}
\alias{rotate_obj_angles}
\title{rotate_obj_angles}
\usage{
rotate_obj_angles(obj, angles)
}
\arguments{
\item{obj}{list of parts and centers}

\item{angles}{integer values for angles of rotation 1 to 360}
}
\description{
rotate_obj_angles
}
